fix file system

+++ scarabei-red/src/com/jfixby/scarabei/red/filesystem/LocalRedFile.java
@@ -2,29 +2,87 @@
 package com.jfixby.scarabei.red.filesystem;
 
 import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 import com.jfixby.scarabei.api.collections.Collections;
 import com.jfixby.scarabei.api.collections.List;
 import com.jfixby.scarabei.api.err.Err;
 import com.jfixby.scarabei.api.file.File;
 import com.jfixby.scarabei.api.file.FileSystem;
+import com.jfixby.scarabei.api.file.LocalFile;
 import com.jfixby.scarabei.api.log.L;
 import com.jfixby.scarabei.api.util.path.AbsolutePath;
+import com.jfixby.scarabei.api.util.path.RelativePath;
 
-public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> extends AbstractRedFile {
-	public abstract String getAbsolutePathString ();
+public final class LocalRedFile extends AbstractRedFile implements LocalFile {
 
 	AbsolutePath<FileSystem> absolute_path;
-	private final T fs;
+	private final LocalFileSystem fs;
 
-	public AbstractLocalFile (final AbsolutePath<FileSystem> absolute_path, final T fileSystem) {
+	final public String getAbsolutePathString (final File file) {
+
+// final Path path = new Path(file.toJavaFile());
+
+// final String mount_point_path_string = "";
+// String relative = toNativePathString(this.getAbsoluteFilePath().getRelativePath().getPathString());
+// if (relative.length() > 0) {
+// relative = UnixFileSystem.OS_SEPARATOR + relative;
+// }
+// final String result = mount_point_path_string + relative;
+// if (result.equals("")) {
+// return "/";
+// }
+// return result;
+//
+// Err.throwNotImplementedYet();
+		return null;
+	}
+
+	@Override
+	final public java.io.File toJavaFile () {
+
+		final RelativePath relative = this.getAbsoluteFilePath().getRelativePath();
+// L.d("relative", relative.toString());
+// final Collection<String> steps = relative.steps();
+// final String unixPath = JUtils.wrapSequence(steps, steps.size(), "", "", "/");
+// L.d("unixPath", unixPath);
+
+// final java.io.File root = path.toFile().getAbsoluteFile();
+// L.d("root", root);
+//
+// L.d("file", this);
+// L.d("path", path);
+//
+// path = path.resolve("");
+///
+		Path path = Paths.get("root").toAbsolutePath().getRoot().toAbsolutePath();
+		for (final String step : relative.steps()) {
+			path = path.resolve(step);
+// L.d("+" + step, path);
+		}
+// L.d("root",);
+// L.d("path", path.toFile().getAbsolutePath());
+// Sys.exit();
+		final java.io.File f = path.toFile().getAbsoluteFile();
+
+// L.d("path", path);
+// L.d("f", f);
+// L.d();
+//
+// Debug.printCallStack();
+// Sys.exit();
+		return f;
+	}
+
+	public LocalRedFile (final AbsolutePath<FileSystem> absolute_path, final LocalFileSystem fileSystem) {
 		this.absolute_path = absolute_path;
 		this.fs = fileSystem;
 
 	}
 
 	@Override
-	final public T getFileSystem () {
+	final public LocalFileSystem getFileSystem () {
 		return this.fs;
 	}
 
@@ -35,7 +93,7 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 
 	@Override
 	final public boolean rename (final String new_name) {
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
+		final java.io.File f = this.toJavaFile();
 		final AbsolutePath<FileSystem> newPath = this.getAbsoluteFilePath().parent().child(new_name);
 		final File newFile = this.getFileSystem().newFile(newPath);
 		final boolean success = f.renameTo(newFile.toJavaFile());
@@ -47,13 +105,13 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 
 	@Override
 	final public boolean isFile () {
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
+		final java.io.File f = this.toJavaFile();
 		return f.isFile();
 	}
 
 	@Override
 	final public long lastModified () {
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
+		final java.io.File f = this.toJavaFile();
 		return f.lastModified();
 	}
 
@@ -65,7 +123,7 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 
 	@Override
 	final public String nameWithoutExtension () {
-		final java.io.File file = new java.io.File(this.getAbsolutePathString());
+		final java.io.File file = this.toJavaFile();
 		final String name = file.getName();
 		final int dotIndex = name.lastIndexOf('.');
 		if (dotIndex == -1) {
@@ -85,34 +143,36 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 
 	@Override
 	final public boolean makeFolder () {
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
+		final java.io.File f = this.toJavaFile();
 		return f.mkdirs();
 	}
 
 	@Override
 	final public boolean delete () throws IOException {
+		final java.io.File f = this.toJavaFile();
+		if (this.getFileSystem().deleteSwitchIsInSafePosition()) {
+			L.e("delete switch is in safe position", this.fs);
+			L.e("ignore file", f);
+			return false;
+		}
+
 		if (this.isFolder()) {
 			this.clearFolder();
 		}
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
-		return f.delete();
-	}
 
-	@Override
-	final public java.io.File toJavaFile () {
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
-		return f;
+		return f.delete();
+
 	}
 
 	@Override
 	final public boolean isFolder () {
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
+		final java.io.File f = this.toJavaFile();
 		return f.isDirectory();
 	}
 
 	@Override
 	final public boolean exists () {
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
+		final java.io.File f = this.toJavaFile();
 		return f.exists();
 	}
 
@@ -121,7 +181,7 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 
 	@Override
 	final public RedFilesList listDirectChildren () {
-		final List<AbstractLocalFile<T>> filesQueue = Collections.newList();
+		final List<LocalRedFile> filesQueue = Collections.newList();
 		filesQueue.add(this);
 		final RedFilesList result = new RedFilesList();
 		collectChildren(filesQueue, result, DIRECT_CHILDREN);
@@ -132,7 +192,7 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 
 	@Override
 	final public RedFilesList listAllChildren () {
-		final List<AbstractLocalFile<T>> filesQueue = Collections.newList();
+		final List<LocalRedFile> filesQueue = Collections.newList();
 		filesQueue.add(this);
 		final RedFilesList result = new RedFilesList();
 		collectChildren(filesQueue, result, ALL_CHILDREN);
@@ -141,16 +201,20 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 
 	}
 
-	static private <T extends AbstractLocalFileSystem> void collectChildren (final List<AbstractLocalFile<T>> filesQueue,
+	static private <T extends LocalFileSystem> void collectChildren (final List<LocalRedFile> filesQueue,
 		final RedFilesList result, final boolean directFlag) {
 		while (filesQueue.size() > 0) {
-			final AbstractLocalFile<T> nextfile = filesQueue.removeElementAt(0);
-			final java.io.File file = new java.io.File(nextfile.getAbsolutePathString());
+			final LocalRedFile nextfile = filesQueue.removeElementAt(0);
+// L.d("absPathString", absPathString);
+			final java.io.File file = nextfile.toJavaFile();
 			if (!file.exists()) {
-				Err.reportError("File does not exist: " + nextfile.getAbsolutePathString());
+				Err.reportError("File does not exist: " + file);
 			}
 
 			if (file.isDirectory()) {
+// L.d("file", file.getAbsolutePath());
+// Sys.exit();
+
 				final String[] list = file.list();
 				if (list == null) {
 					L.e("list() is null", file);
@@ -162,7 +226,7 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 					final String file_i = files.getElementAt(i);
 					final AbsolutePath<FileSystem> absolute_file = nextfile.absolute_path.child(file_i);
 					final T fs = (T)absolute_file.getMountPoint();
-					final AbstractLocalFile<T> child = (AbstractLocalFile<T>)fs.newFile(absolute_file);
+					final LocalRedFile child = fs.newFile(absolute_file);
 					result.add(child);
 					if (directFlag == ALL_CHILDREN) {
 
@@ -181,18 +245,22 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 		}
 	}
 
-	final public String toAbsolutePathString () {
-		return this.getAbsolutePathString();
-	}
+// final public String getAbsolutePathString () {
+// return this.fs.getAbsolutePathString(this);
+// }
 
-	final public java.io.File getJavaFile () {
-		final java.io.File file = new java.io.File(this.getAbsolutePathString());
-		return file;
-	}
+// final public String toAbsolutePathString () {
+// return this.getAbsolutePathString();
+// }
+
+// final public java.io.File getJavaFile () {
+// final java.io.File file = new java.io.File(this.getAbsolutePathString());
+// return file;
+// }
 
 	@Override
 	final public long getSize () {
-		final java.io.File file = new java.io.File(this.getAbsolutePathString());
+		final java.io.File file = this.toJavaFile();
 		if (file.isFile()) {
 			return file.length();
 		} else {
@@ -202,38 +270,8 @@ public abstract class AbstractLocalFile<T extends AbstractLocalFileSystem> exten
 
 	@Override
 	final public String getName () {
-		final java.io.File f = new java.io.File(this.getAbsolutePathString());
+		final java.io.File f = this.toJavaFile();
 		return f.getName();
-	}
-
-	@Override
-	final public int hashCode () {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((this.getAbsolutePathString() == null) ? 0 : this.getAbsolutePathString().hashCode());
-		return result;
-	}
-
-	@Override
-	final public boolean equals (final Object obj) {
-		if (this == obj) {
-			return true;
-		}
-		if (obj == null) {
-			return false;
-		}
-		if (this.getClass() != obj.getClass()) {
-			return false;
-		}
-		final AbstractLocalFile other = (AbstractLocalFile)obj;
-		if (this.getAbsolutePathString() == null) {
-			if (other.getAbsolutePathString() != null) {
-				return false;
-			}
-		} else if (!this.getAbsolutePathString().equals(other.getAbsolutePathString())) {
-			return false;
-		}
-		return true;
 	}
 
 }
